+++ routes/index.js
@@ -4,6 +4,8 @@ module.exports = {
   'get /agolworkers': 'getQueueCounts',
   'get /agol/:id': 'find',
   'get /agol/:id/*': 'setHostKey',
+  'put /agol/:id/*': 'setHostKey',
+  'post /agol/:id/*': 'setHostKey',
   'delete /agol/:id': 'del',
   'delete /agol/:id/:item/:layer': 'deleteItemData',
   'get /agol/:id/:item/:layer/geohash': 'getGeohash',
@@ -21,5 +23,8 @@ module.exports = {
   'get /agol/:id/:item/FeatureServer/:layer': 'featureserver',
   'get /agol/:id/:item/FeatureServer/:layer/:method': 'featureserver',
   'post /agol/:id/:item/FeatureServer/:layer/:method': 'featureserver',
-  'get /agol/:id/:item/FeatureServer/:layer/geohash': 'getGeohash'
+  'get /agol/:id/:item/FeatureServer/:layer/geohash': 'getGeohash',
+  'get /agol/:id/:item/:layer/expiration': 'getExpiration',
+  'put /agol/:id/:item/:layer/expiration': 'setExpiration',
+  'post /agol/:id/:item/:layer/expiration': 'setExpiration'
 }